requirement.txt
pandas-datareader==0.9.0
numpy==1.22.3
sklearn==0.23.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.10.0
tensorflow-intel==2.10.0
tensorflow-io-gcs-filesystem==0.25.0
tensorflow-metadata==1.12.0
mpmath==1.1.0
matplotlib==3.5.2
pathos==0.3.2
keras==2.10.0
yfinance==0.1.62
keras-self-attention==0.51.0



